♻️ refactor(page-share): extract to business slot stubs
