+++ src/business/client/features/PageShare/PageShareLayout.tsx
@@ -1,0 +1,5 @@
+import { type ReactNode } from 'react';
+
+export default function PageShareLayout({ children }: { children: ReactNode; pageId?: string }) {
+  return children;
+}
