+++ src/business/client/features/PageShare/ReadOnlyPageViewer.tsx
@@ -1,0 +1,5 @@
+import { type SharedDocumentData } from '@lobechat/types';
+
+export default function ReadOnlyPageViewer(_: { data: SharedDocumentData }) {
+  return null;
+}
