+++ src/business/client/features/PageShare/ShareButton.tsx
@@ -1,0 +1,3 @@
+export default function ShareButton(_: { documentId: string }) {
+  return null;
+}
