+++ src/business/server/lambda-routers/pageShare.ts
@@ -1,0 +1,3 @@
+import { router } from '@/libs/trpc/lambda';
+
+export const pageShareRouter = router({});
