+++ src/routes/(main)/_layout/index.tsx
@@ -9,6 +9,7 @@ import { lazy, Suspense } from 'react';
 import { HotkeysProvider } from 'react-hotkeys-hook';
 import { Outlet, useLocation, useParams } from 'react-router-dom';
 
+import PageShareLayout from '@/business/client/features/PageShare/PageShareLayout';
 import Loading from '@/components/Loading/BrandTextLoading';
 import { isDesktop } from '@/const/version';
 import { BANNER_HEIGHT } from '@/features/AlertBanner/CloudBanner';
@@ -23,11 +24,9 @@ import TabCacheBridges from '@/features/Electron/titlebar/TabBar/TabCacheBridges
 import TitleBar from '@/features/Electron/titlebar/TitleBar';
 import HotkeyHelperPanel from '@/features/HotkeyHelperPanel';
 import NavPanel from '@/features/NavPanel';
-import PublishedShell from '@/features/PageShare/PublishedShell';
 import { RouteMetaBridge } from '@/features/RouteMeta';
 import { useFeedbackModal } from '@/hooks/useFeedbackModal';
 import { usePlatform } from '@/hooks/usePlatform';
-import { useSharedPageProbe } from '@/hooks/useSharedPageProbe';
 import { MarketAuthProvider } from '@/layout/AuthProvider/MarketAuth';
 import CmdkLazy from '@/layout/GlobalProvider/CmdkLazy';
 import dynamic from '@/libs/next/dynamic';
@@ -62,76 +61,67 @@ const Layout: FC = () => {
   const params = useParams<{ id?: string }>();
   const onPageRoute = isPageRoute(location.pathname);
   const pageId = onPageRoute && params.id ? getIdFromIdentifier(params.id, 'docs') : undefined;
-  const { data: probe, error: probeError } = useSharedPageProbe(pageId);
-
-  const isGuestPageRoute = !!pageId && (!!probeError || (probe ? !probe.isOwner : false));
-
-  if (isGuestPageRoute) {
-    return (
-      <PublishedShell data={probe} error={probeError}>
-        <Outlet context={{ error: probeError, probe }} />
-      </PublishedShell>
-    );
-  }
 
   return (
-    <HotkeysProvider initiallyActiveScopes={[HotkeyScopeEnum.Global]}>
-      <RouteMetaBridge />
-      {isDesktop && <TabCacheBridges />}
-      <Suspense fallback={null}>
-        {isDesktop && <DesktopAutoOidcOnFirstOpen />}
-        {isDesktop && <DesktopNavigationBridge />}
-        {isDesktop && <DesktopFileMenuBridge />}
-        {isDesktop && <OverlaySnapshotPublisher />}
-        {isDesktop && <OverlayCaptureUploader />}
-        {isDesktop && <OverlayMessageDispatcher />}
-        {showCloudPromotion && <CloudBanner />}
-      </Suspense>
-      {isDesktop && <AuthRequiredModal />}
-      {isDesktop && <ZoomHUD />}
+    <PageShareLayout pageId={pageId}>
+      <HotkeysProvider initiallyActiveScopes={[HotkeyScopeEnum.Global]}>
+        <RouteMetaBridge />
+        {isDesktop && <TabCacheBridges />}
+        <Suspense fallback={null}>
+          {isDesktop && <DesktopAutoOidcOnFirstOpen />}
+          {isDesktop && <DesktopNavigationBridge />}
+          {isDesktop && <DesktopFileMenuBridge />}
+          {isDesktop && <OverlaySnapshotPublisher />}
+          {isDesktop && <OverlayCaptureUploader />}
+          {isDesktop && <OverlayMessageDispatcher />}
+          {showCloudPromotion && <CloudBanner />}
+        </Suspense>
+        {isDesktop && <AuthRequiredModal />}
+        {isDesktop && <ZoomHUD />}
 
-      <Suspense fallback={null}>{isDesktop && <TitleBar />}</Suspense>
-      <DndContextWrapper>
-        <Flexbox
-          horizontal
-          className={cx(isPWA ? styles.mainContainerPWA : styles.mainContainer)}
-          width={'100%'}
-          height={
-            isDesktop
-              ? `calc(100% - ${TITLE_BAR_HEIGHT}px)`
-              : showCloudPromotion
-                ? `calc(100% - ${BANNER_HEIGHT}px)`
-                : '100%'
-          }
-        >
-          <NavPanel />
-          <DesktopLayoutContainer>
-            <MarketAuthProvider isDesktop={isDesktop}>
-              <DesktopHomeLayout>
-                <DesktopHome />
-              </DesktopHomeLayout>
-              <Suspense fallback={<Loading debugId="DesktopMainLayout > Outlet" />}>
-                <Outlet context={{ probe }} />
-              </Suspense>
-            </MarketAuthProvider>
-          </DesktopLayoutContainer>
-        </Flexbox>
-      </DndContextWrapper>
-      <Suspense fallback={null}>
-        <HotkeyHelperPanel />
-        <RegisterHotkeys />
-        <CmdkLazy />
-        {isFeedbackModalOpen && (
-          <Suspense fallback={null}>
-            <FeedbackModal
-              initialValues={feedbackInitialValues}
-              open={isFeedbackModalOpen}
-              onClose={closeFeedbackModal}
-            />
-          </Suspense>
-        )}
-      </Suspense>
-    </HotkeysProvider>
+        <Suspense fallback={null}>{isDesktop && <TitleBar />}</Suspense>
+        <DndContextWrapper>
+          <Flexbox
+            horizontal
+            className={cx(isPWA ? styles.mainContainerPWA : styles.mainContainer)}
+            width={'100%'}
+            height={
+              isDesktop
+                ? `calc(100% - ${TITLE_BAR_HEIGHT}px)`
+                : showCloudPromotion
+                  ? `calc(100% - ${BANNER_HEIGHT}px)`
+                  : '100%'
+            }
+          >
+            <NavPanel />
+            <DesktopLayoutContainer>
+              <MarketAuthProvider isDesktop={isDesktop}>
+                <DesktopHomeLayout>
+                  <DesktopHome />
+                </DesktopHomeLayout>
+                <Suspense fallback={<Loading debugId="DesktopMainLayout > Outlet" />}>
+                  <Outlet />
+                </Suspense>
+              </MarketAuthProvider>
+            </DesktopLayoutContainer>
+          </Flexbox>
+        </DndContextWrapper>
+        <Suspense fallback={null}>
+          <HotkeyHelperPanel />
+          <RegisterHotkeys />
+          <CmdkLazy />
+          {isFeedbackModalOpen && (
+            <Suspense fallback={null}>
+              <FeedbackModal
+                initialValues={feedbackInitialValues}
+                open={isFeedbackModalOpen}
+                onClose={closeFeedbackModal}
+              />
+            </Suspense>
+          )}
+        </Suspense>
+      </HotkeysProvider>
+    </PageShareLayout>
   );
 };
