+++ src/server/routers/lambda/index.ts
@@ -2,6 +2,7 @@
  * This file contains the root router of Lobe Chat tRPC-backend
  */
 import { accountDeletionRouter } from '@/business/server/lambda-routers/accountDeletion';
+import { pageShareRouter } from '@/business/server/lambda-routers/pageShare';
 import { referralRouter } from '@/business/server/lambda-routers/referral';
 import { spendRouter } from '@/business/server/lambda-routers/spend';
 import { subscriptionRouter } from '@/business/server/lambda-routers/subscription';
@@ -129,6 +130,7 @@ export const lambdaRouter = router({
   video: videoRouter,
   webBrowsing: webBrowsingRouter,
   accountDeletion: accountDeletionRouter,
+  pageShare: pageShareRouter,
   referral: referralRouter,
   spend: spendRouter,
   subscription: subscriptionRouter,
